Diff screens desgin changes

+++ app/src/main/java/com/example/services/adapters/cart/CartListAdapter.kt
@@ -49,7 +49,7 @@ class CartListAdapter(
     override fun onBindViewHolder(@NonNull holder: ViewHolder, position: Int) {
         viewHolder = holder
         holder.binding!!.tvCatName.text = addressList[position].service?.name
-        holder.binding!!.tvQuantity.setText( addressList[position].quantity)
+        holder.binding!!.tvQuantity.setText(addressList[position].quantity)
         holder.binding!!.tvOfferPrice.setText(
             GlobalConstants.Currency + "" + addressList[position].price.toString()
         )
@@ -71,14 +71,14 @@ class CartListAdapter(
         }
 */
         holder.binding!!.imgPlus.setOnClickListener {
-            if ( addressList[position].quantity!!.toInt() <= 5) {
+            if (addressList[position].quantity!!.toInt() <= 5) {
                 var quantity = addressList[position].quantity!!.toInt()
                 quantity++
                 // serviceDetailBinding.btnSubmit.isEnabled = false
                 holder.binding!!.tvQuantity.setText(quantity.toString())
                 //   serviceDetailBinding.btnSubmit.visibility = View.VISIBLE
                 //callGetTimeSlotsApi()
-                var price = quantity * addressList[position].price!!.toInt()
+                var price = quantity * addressList[position].price!!.toDouble()
                 //  tvTotalPrice?.setText(GlobalConstants.Currency + " " + price.toString())
                 mContext.clickAddButton(position, price, quantity)
             }
@@ -88,10 +88,10 @@ class CartListAdapter(
             if (addressList[position].quantity!!.toInt() > 1) {
                 var quantity = addressList[position].quantity!!.toInt()
                 quantity--
-                var price = quantity * addressList[position].price!!.toInt()
+                var price = quantity * addressList[position].price!!.toDouble()
                 // tvTotalPrice?.setText(GlobalConstants.Currency + " " + price.toString())
                 //callGetTimeSlotsApi()
-                mContext.clickMinusButton(position,price,quantity)
+                mContext.clickMinusButton(position, price, quantity)
                 holder.binding!!.tvQuantity?.setText(quantity.toString())
             }
             if (addressList[position].quantity!!.toInt() == 1) {
